requirements.txt
Django==5.1.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.8.0
drf-yasg==1.21.7
oauthlib==3.2.2
requests==2.31.0
social-auth-app-django==5.4.0
django-allauth==0.59.0
random-username==1.0.2
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
whitenoise==6.9.0
